Replace `Linter::Javscript` with `hound-jshint`
- Remove `jshint` gem.
- Remove `jshint` from `BETA_LANGUAGES` list.
- Move `config/style_guides/javascript.json` to [hound-jshint].

[hound-jshint][https://github.com/thoughtbot/hound-jshint/master/config/.jshintrc]

https://trello.com/c/NDWEJI1G
