+++ Gemfile
@@ -19,7 +19,6 @@ gem "haml-rails"
 gem "haml_lint"
 gem "high_voltage"
 gem "jquery-rails", "~> 3.1.3"
-gem "jshintrb"
 gem "neat"
 gem "octokit"
 gem "omniauth-github"
