+++ app/models/hound_config.rb
@@ -3,7 +3,6 @@ class HoundConfig
   BETA_LANGUAGES = %w(
     eslint
     jscs
-    jshint
     mdast
     python
     swift
