--- config/style_guides/javascript.json
@@ -1,33 +1,0 @@
-{
-  "asi": false,
-  "bitwise": true,
-  "browser": true,
-  "camelcase": true,
-  "curly": true,
-  "forin": true,
-  "immed": true,
-  "latedef": "nofunc",
-  "maxlen": 80,
-  "newcap": true,
-  "noarg": true,
-  "noempty": true,
-  "nonew": true,
-  "predef": [
-    "$",
-    "jQuery",
-
-    "jasmine",
-    "beforeEach",
-    "describe",
-    "expect",
-    "it",
-
-    "angular",
-    "inject",
-    "module"
-  ],
-  "quotmark": true,
-  "trailing": true,
-  "undef": true,
-  "unused": true
-}
